+++ maybetype/errors.py
@@ -1,0 +1,2 @@
+class ResultUnwrapError(Exception):
+    """Raised when calling ``Result.unwrap()`` if the instance is not ``Ok``."""